Avances Fb

+++ src/Loogares/UsuarioBundle/Security/User/FacebookProvider.php
@@ -133,9 +133,10 @@ class FacebookProvider implements UserProviderInterface
                             $fn = new LoogaresFunctions();
                             $filename = $fn->generarSlug($user->getNombre().'-'.$user->getApellido().'-'.$user->getId());
                             $user->setImagenFull($filename.'.jpg');  
-                            file_put_contents('assets/images/usuarios/'.$filename, $result);
-                            $em->flush();
-                            unlink('assets/images/temp/'.$fn);
+                            if(file_put_contents('assets/images/usuarios/'.$filename.'.jpg', $result)) {
+                                $em->flush();
+                                unlink('assets/images/temp/'.$fn);
+                            }                            
                         }                        
                     }
                 }
